add comments to debug the openTK error

+++ src/Shaders/Dependencies/WidthWindowDependence.cs
@@ -17,5 +17,5 @@ public class WidthWindowDependence : ShaderDependence
         => sb.AppendLine("uniform float width;");
 
     public override Action AddOperation(ShaderManager ctx)
-        => () => ctx.SetFloat("width", (float)Window.Width);
+        => () => ctx.SetFloat("width", Window.Width);
 }
